#439 moved the free call related details to the free call api from payment channel api
